+++ package.json
@@ -20,6 +20,7 @@
   "devDependencies": {
     "@swc-jotai/react-refresh": "^0.1.0",
     "@types/howler": "^2.2.10",
+    "@types/lodash": "^4.14.200",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
